FBCanalysis: A package for developing and evaluating biomedical time series data clustering models based on Fluctuation Based Clustering (FBC)

This R package aims to offer researchers with fast tools for clustering patient time series data and confirming the distinction using additional metrics such as population parameter enrichment analysis, stability after random data removal, and conventional cluster stability measures. The package attempts to conveniently apply computational methods and capabilities for developing and evaluating unsupervised clustering models with the goal of data-drivenly categorizing asthmatic patients according to their illness dynamics.

Details

clustering may be used within the proposed package to identify significant diverse groupings in a patient population, and enrichment analysis is used to examine any possible correlations with clinically relevant characteristics.

The R package thus aims to offer researchers with fast tools for clustering patient time series data and confirming the distinction using additional metrics such as population parameter enrichment analysis, stability after random data removal, and conventional cluster stability measures

Time series data preparation and visualization functions

patient_list

patient_ts_plot

patient_boxplot

patient_hist

Earth Mover's Distance processing functions

emd_matrix

emd_heatmap

max_fluc

Clustering techniques

clust_matrix

Enrichment analysis functions

add_enrich

add_clust2enrich

add_clust2ts

enr_obs_clust

sim_sample_enr

Determine cluster stability upon random data removal

sim_jaccard_global

sim_jaccard_emd

sim_jaccard_emd_2

jaccard_run_global

jaccard_run_emd

jaccard_run_emd_2

reap_freq

reap_freq_run

Cluster stability measure validation

init_clValid

clValid_flow

Helper function

znorm

See also