Add clustering assignments from clustering output to preprocessed enrichment data frame.

Add clustering assignments from clustering output to preprocessed enrichment data frame.

Usage

add_clust2enrich(enrich, clustdat)

Arguments

enrich

Preprocessed enrichment data frame (also see function: add_enrich)

clustdat

Object of type list storing clustering data (also see function: clust_matrix)

Value

Processed enrichment data frame with added column indicating cluster assignments

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from top quantile 	 (1) 
#>  Sample missing values with bottom quantile 	 (2) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression (3) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression (4) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization (5) 
#> 
#>  Interpolate missing values by Linear Spline (6) 
#>  Interpolate missing values by Cubic C2 Spline (7) 
#> 
#> Select measure for filling up NA values: 
#> Error in `[.data.frame`(raw.data, , "Patient_ID"): undefined columns selected
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
#> Error in as.vector(x, mode): cannot coerce type 'closure' to vector of type 'any'
enr <- add_enrich(list,
'https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/enrich/enrichment.csv')
#> 
#> 
#>      Patient_ID Gender    Location Age BMI
#> 1          b301      F        City  10  26
#> 2 patient_ID_32      F        City  12  27
#> 3     testpat_2      F        City  16  NA
#> 4          a100      F Countryside   8  20
#> 5          a300      M Countryside   7  22
#> 
#> 
#>          V1     V2       V3  V4  V5
#>  Patient_ID Gender Location Age BMI
#>           1      2        3   4   5
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
enr <- add_clust2enrich(enr, clustering)
#> Error in as.data.frame(clustdat$Cls): object 'clustering' not found