Interactive console workflow to calculate and evaluate cluster validation measures which have been determined previously by the call init_clValid.

Interactive console workflow to calculate and evaluate cluster validation measures which have been determined previously by the call init_clValid.

Usage

clValid_flow(matrix, par)

Arguments

matrix

Earth Mover's Distance Matrix for processed patient time series data (also see functions: emd_matrix, patient_list)

par

Object of type list storing clustering methods and cluster range of interest; initialized via function: init_clValid

Value

Object of type list storing chosen clustering method and number of clusters (can be then used for function clust_matrix)

Details

The call guides through an interactive workflow and generates cluster evaluation measures, stores and lists, visualizes corresponding plots and lets the user decide which technique is the prefered one. Once the user has chosen his favourite, the flow continues to the function clust_matrix and generates the respective clustering output. The internal cluster validation methods utilize just the dataset and the clustering partition as input and evaluates the clustering’s quality by using intrinsic information included in the data.

The call calculates Connectivity, Silhouette width and Dunn index. Connectivity describes the connectness to neighbors of particular clustering partition and should be minimized. Silhouette width defines the average silhouette value for each observation and should be maximized. The Dunnn index is a definition for Ratio of shortest distance between non-cluster observation and greatest intra-cluster distance and should be maximized likewise.

Furthermore, cluster stability measures are available, namely Average proportion of non-overlap (APN), Average distance (AD), Average distance between means (ADM) and Figure of merit (FOM). APN is the average proportion of observations that are not clustered using complete and leaky data. AD defines the average distance in observations for both complete and leaky data. ADM deals with the average distance between cluster centers in complete and leaky data. FOM is a measure for average intra- cluster variance in leaky data. All measures should be minimized. Furthermore, Rank Aggregation may be performed. It approaches to provide a generic and flexible framework for objectively integrating several ordered lists in a suitable and efficient way. The used technique for evaluating clustering the rank is by a Cross-entropy approach, which is incorporating Spearman’s footrule distance measure. In the end a recommendation for the best fitting clustering model is given.

References

Guy Brock, Vasyl Pihur, Susmita Datta, and Somnath Datta. clvalid: An r package for cluster validation. Journal of Statistical Software, 25:1–22, 2008.

Julia Handl, Joshua Knowles, and Douglas B Kell. Computational cluster validation in post-genomic data analysis. Bioinformatics, 21(15):3201–3212, 2005.

Peter J Rousseeuw. Silhouettes: a graphical aid to the interpretation and validation of cluster analysis. Journal of computational and applied mathematics, 20:53–65, 1987.

Joseph C Dunn. Well-separated clusters and optimal fuzzy partitions. Journal of cybernetics, 4(1):95–104, 1974.

Vasyl Pihur, Susmita Datta, and Somnath Datta. Weighted rank aggregation of cluster validation measures: a monte carlo cross-entropy approach. Bioinformatics, 23(13):1607–1615, 2007.

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from top quantile 	 (1) 
#>  Sample missing values with bottom quantile 	 (2) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression (3) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression (4) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization (5) 
#> 
#>  Interpolate missing values by Linear Spline (6) 
#>  Interpolate missing values by Cubic C2 Spline (7) 
#> 
#> Select measure for filling up NA values: 
#> Error in `[.data.frame`(raw.data, , "Patient_ID"): undefined columns selected
#Sampling frequency is supposed to be daily
distmat <- emd_matrix(list, "PEF", maxIter = 5000)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Error in plist[[i]]: object of type 'builtin' is not subsettable
parameters <- init_clValid()
#> Error in data.frame(..., check.names = FALSE): arguments imply differing number of rows: 10, 9
output <- clValid_flow(distmat, parameters)
#> Do you want to perform internal cluster validation (y/n)?: 
#> Do you want to perform cluster stability validation (y/n)?: 
#> Do you want to perform rank aggregation? (y/n): 
#> 
#> Error in clValid_flow(distmat, parameters): object 'distmat' not found