Generate Earth Mover's Distance Matrix

Generate an Earth Mover's Distance Matrix for time series data distributions pairs out of a preprocessed time series data list.

emd_matrix(plist, parameter, maxIter, normalize)

Arguments

plist

List storing patient time series data (also see function: patient_list)

parameter

Parameter of interest to determine Earth Mover's Distances between distributions

maxIter

Maximum of iterations to calculate Earth Mover's Distance (default: 500)

normalize

Indicates if parameters delivered needs to be normalized or not (TRUE by default)

Value

Earth Mover's Distance Square Matrix of type matrix

Details

the function may compute the EMD for each Patient ID pairi,j using the normalized distributions. EMD is a distance measure between two probability distributions over a region D. Informally, if the distributions are viewed as two distinct methods of accumulating a certain quantity of earth gravel across the region D. EMD is the smallest cost associated with converting one pile to another, where the cost is supposed to equal the quantity of material transferred multiplied by the distance traveled. A unit of labor is defined in this context as conveying a unit of earth across a unit of ground distance. A distribution may be described as a collection of clusters, each of which is defined by its mean or mode and the proportion of the distribution that belongs to it. This representation is referred to as the distribution’s signature. Both signatures may be of varying sizes. Simple distributions, for example, have lower signatures than complex distributions. (also see emd for further details)

References

Yossi Rubner, Carlo Tomasi, and Leonidas J Guibas. A metric for distributions with applications to image databases. In Sixth International Conference on Computer Vi- sion (IEEE Cat. No. 98CH36271), pages 59–66. IEEE, 1998.

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from First quartile 	 (1) 
#>  Sample missing values from Third quartile 	 (2) 
#> 
#>  Sample missing values from 5th percentile 	 (3) 
#>  Sample missing values from 95th percentile 	 (4) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression 		 (5) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression 		 (6) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization 	 (7) 
#> 
#>  Interpolate missing values by Linear Spline 	 (8) 
#>  Interpolate missing values by Cubic C2 Spline 	 (9) 
#> 
#> Select measure for filling up NA values: 
#> Error in patient_list("https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",     GitHub = TRUE): object 'answer_a' not found
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Error in plist[[i]]: object of type 'builtin' is not subsettable