Observe a specific cluster of interest on preporcessed enrichment and time series data for overview and p-values. The p-values are calculated by a Wilcox Rank Sum test for continuous data and a Fisher's Exact Test for categorical enrichment data.

Observe a specific cluster of interest on preporcessed enrichment and time series data for overview and p-values. The p-values are calculated by a Wilcox Rank Sum test for continuous data and a Fisher's Exact Test for categorical enrichment data.

Usage

enr_obs_clust(ts.dat, enrich, clustno)

Arguments

ts.dat

Processed data frame storing time series data and cluster assignments (also see function: add_clust2ts)

enrich

Processed data frame storing enrichment data and cluster assignments (also see function: add_clust2enrich)

clustno

Cluster number of interest

Value

Terminal output presenting summary of time series and enrichment data with corresponding p-values

Details

There are two techniques to compute the p-value for continuous or categorical data, according to past work. In order to determine the relevant p-value inside a cluster of interest, the data distribution within the cluster should be compared to the data distribution outside the cluster. Prior to conducting the related probability tests, one data processing step is performed, namely the construction of two data distributions, one including only data included inside the cluster and another comprising data from outside the cluster, for the purpose of comparing them.

The Mann-Whitney Test, sometimes referred to as the Wilcoxon rank-sum test (WRS), is used to measure the significance of continuous variables within the observed distribution. The WRS is used to test if the central tendency of two independent samples is different. When the t-test for independent samples does not meet the requirements, the WRS is used. The null hypothesis H0 states that the populations’ distributions are equal. H1 is the alternative hypothesis meaning that the distributions are not equal. The test is consistent under the broader formulation only when the following happens under H1.

The hypergeometric test or Fisher’s exact test (FET) is used to analyse categorical variables within the enriched data set. It is a statistical significance test for contingency tables that is employed in the study of them. The test is helpful for categorical data derived from object classification. It is used to assess the importance of associations and inconsistencies between classes. The FET is often used in conjunction with a 2 × 2 contingency table that represents two categories for a variable, as well as assignment inside or outside of the cluster. The p-value is calculated as if the table’s margins are fixed. This results in a hypergeometric distribution of the numbers in the table cells under the null hypothesis of independence. A hypergeometric distribution is a discrete probability distribution that describes the probability of k successes, defined as random draws for which the object drawn has a specified feature in n draws without replacement from a finite population of size N containing exactly K objects with that feature, where each draw is either successful or unsuccessful. The test is only practicable for normal computations in the presence of a 2 × 2 contingency table. However, the test’s idea may be extended to the situation of a m × n table in general. Statistics programs provide a Monte Carlo approach for approximating the more general case.

References

Siegel Sidney. Nonparametric statistics for the behavioral sciences. The Journal of Nervous and Mental Disease, 125(3):497, 1957.

Kinley Larntz. Small-sample comparisons of exact levels for chi-squared goodness-of-fit statistics. Journal of the American Statistical Association, 73(362):253–263, 1978.

Cyrus R Mehta and Nitin R Patel. A network algorithm for performing fisher’s exact test in r× c contingency tables. Journal of the American Statistical Association, 78(382):427–434, 1983.

Aravind Subramanian, Pablo Tamayo, Vamsi K Mootha, Sayan Mukherjee, Benjamin L Ebert, Michael A Gillette, Amanda Paulovich, Scott L Pomeroy, Todd R Golub, Eric S Lander, et al. Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles. Proceedings of the National Academy of Sciences, 102(43):15545–15550, 2005.

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from top quantile 	 (1) 
#>  Sample missing values with bottom quantile 	 (2) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression (3) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression (4) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization (5) 
#> 
#>  Interpolate missing values by Linear Spline (6) 
#>  Interpolate missing values by Cubic C2 Spline (7) 
#> 
#> Select measure for filling up NA values: 
#> Error in `[.data.frame`(raw.data, , "Patient_ID"): undefined columns selected
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
#> Error in as.vector(x, mode): cannot coerce type 'closure' to vector of type 'any'
enr <- add_enrich(list,
'https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/enrich/enrichment.csv')
#> 
#> 
#>      Patient_ID Gender    Location Age BMI
#> 1          b301      F        City  10  26
#> 2 patient_ID_32      F        City  12  27
#> 3     testpat_2      F        City  16  NA
#> 4          a100      F Countryside   8  20
#> 5          a300      M Countryside   7  22
#> 
#> 
#>          V1     V2       V3  V4  V5
#>  Patient_ID Gender Location Age BMI
#>           1      2        3   4   5
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
enr <- add_clust2enrich(enr, clustering)
#> Error in as.data.frame(clustdat$Cls): object 'clustering' not found
ts <- add_clust2ts(list, clustering)
#> Error in do.call(rbind, plist): second argument must be a list
enr_obs_clust(ts, enr, 1)
#> Error in dplyr::filter(enrich, Cluster == clustno): Problem while computing `..1 = Cluster == clustno`.
#> Caused by error:
#> ! object 'Cluster' not found