Simulate random data removal range via Earth Mover's Distance Cognate Cluster cluster assignment approach

Simulate amount of random data removal from time series data list and determine Jaccard index via Earth Mover's Distance approach for multiple random data removal steps for a specific cluster of interest.

jaccard_run_emd(
  plist,
  parameter,
  n_simu,
  method,
  clust_num,
  n_clust,
  range,
  maxIter,
  normalize
)

Arguments

plist

Object of type list storing patient time series data (also see function: patient_list)

parameter

Parameter of interest in time series data list

n_simu

Number of simulations

method

Clustering method (also see function: clust_matrix)

clust_num

Cluster of interest

n_clust

Number of clusters

range

Range to simulate random data removal (e.g. c(0.1,0.2,0.5,0.7,0.8))

maxIter

Maximum iterations to determine Earth Mover's Distances (also see function: emd_matrix); default is 5,000 for this function

normalize

Indicates if parameter indicated needs to be normalized or not (TRUE by default)

Value

Object of type list storing Jaccard indices for each indicated random data removal step and visualized results in a boxplot

Details

See sim_jaccard_emd for more detailed approach on Jaccard index determination. The difference in this function is that now only one cluster is observed für multiple amounts of random data removal where for each data removal step defined the resulting Jaccard indices are stored in a list object. Furthermore, a boxplot visualization is generated, in the style of recent publications.

References

Yossi Rubner, Carlo Tomasi, and Leonidas J Guibas. A metric for distributions with applications to image databases. In Sixth International Conference on Computer Vision (IEEE Cat. No. 98CH36271), pages 59–66. IEEE, 1998.

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from First quartile 	 (1) 
#>  Sample missing values from Third quartile 	 (2) 
#> 
#>  Sample missing values from 5th percentile 	 (3) 
#>  Sample missing values from 95th percentile 	 (4) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression 		 (5) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression 		 (6) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization 	 (7) 
#> 
#>  Interpolate missing values by Linear Spline 	 (8) 
#>  Interpolate missing values by Cubic C2 Spline 	 (9) 
#> 
#> Select measure for filling up NA values: 
#> Error in patient_list("https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",     GitHub = TRUE): object 'answer_a' not found
#Sampling frequency is supposed to be daily
output <- jaccard_run_emd(list,"PEF",10,"hierarchical",1,3,c(0.005,0.01,0.05,0.1,0.2))
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Error in plist[[i]]: object of type 'builtin' is not subsettable