Determine pair of maximum fluctuation difference

Determine the pair of maximum fluctuation difference on time series data distribution from a preprocessed list or Earth Mover's Distance square matrix.

max_fluc(input, parameter, maxIter, normalize)

Arguments

input

Either a list storing time series data or EMD martrix (also see functions: patient_list, emd_matrix)

parameter

Parameter of interest from time series data list

maxIter

Maximum of iterations to apply for calculation of Earth Mover's Distannce (also see function: emd_matrix)

normalize

Indicates if parameters delivered needs to be normalized or not (TRUE by default)

Value

Console output with Patient_ID pair, corresponding Earth Mover's Distance and visualized boxplot of both time series data distributions

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from First quartile 	 (1) 
#>  Sample missing values from Third quartile 	 (2) 
#> 
#>  Sample missing values from 5th percentile 	 (3) 
#>  Sample missing values from 95th percentile 	 (4) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression 		 (5) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression 		 (6) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization 	 (7) 
#> 
#>  Interpolate missing values by Linear Spline 	 (8) 
#>  Interpolate missing values by Cubic C2 Spline 	 (9) 
#> 
#> Select measure for filling up NA values: 
#> Error in `[.data.frame`(raw.data, , "Patient_ID"): undefined columns selected
#Sampling frequency is supposed to be daily
max_fluc(list, "PEF")
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Error in plist[[i]]: object of type 'builtin' is not subsettable