Determine reappearance frequency for range

Determine the reappearance frequency for clustered elements in perturbed data for specified amounts of random data removal

reap_freq_run(
  plist,
  parameter,
  n_simu,
  method,
  n_clust,
  range,
  Iter,
  normalize
)

Arguments

plist

Object of type list storing patient time series data (also see function: patient_list)

parameter

Parameter of interest in time series data list

n_simu

Number of simulations

method

Clustering method (also see function: clust_matrix)

n_clust

Number of clusters

range

Range to simulate random data removal (e.g. c(0.1,0.2,0.5,0.7,0.8))

Iter

Maximum iterations to determine Earth Mover's Distances (also see function: emd_matrix); default is 5,000 for this function

normalize

Indicates if parameter indicated needs to be normalized or not (TRUE by default)

Value

List storing vectors with reappearance frequencies for each simulation run for each removal amount indicated as well as illustrated results in a boxplot

Details

The procedure on determining the reappearance frequenncy is described in reap_freq.

References

Edgar Delgado-Eckert, Oliver Fuchs, Nitin Kumar, Juha Pekkanen, Jean-Charles Dalphin, Josef Riedler, Roger Lauener, Michael Kabesch, Maciej Kupczyk, Sven-Erik Dahlen, et al. Functional phenotypes determined by fluctuation-based clustering of lung function measurements in healthy and asthmatic cohort participants. Thorax, 73(2):107–115, 2018.