Simulate random data removal via Global Cognate Cluster cluster assignment approach

Simulate random data removal for a removal amount with indicated number of simulations from time series data list and determine Jaccard index for all cluster via Global Cognate Cluster cluster assignment approach

sim_jaccard_global(
  plist,
  parameter,
  removal,
  n_simu,
  method,
  n_clust,
  maxIter,
  normalize
)

Arguments

plist

Object of type list storing patient time series data (also see function: patient_list)

parameter

Parameter of interest in time series data list

removal

Amount of random data removal to determine Jaccard index

n_simu

Number of simulations

method

Clustering method (also see function: clust_matrix)

n_clust

Number of clusters (also see function: clust_matrix)

normalize

Indicates if parameter indicated needs to be normalized or not (TRUE by default)

Iter

Maximum iterations to determine Earth Mover's Distances (also see function: emd_matrix); default is 5,000 for this function

Value

Object of type matrix storing received Jaccard indices for indicated amount of random data removal for all clusters

Details

The cognate cluster approach works in the manner that first a Gold Standard cluster is determined meaning the cluster assignments without any data removal. Subsequently, random data is removed from the original, complete data and clustering is performed again on the leaky data. The cluster determined to be cognate to the Gold Standard cluster is the one with the highest overlap in cluster members, meaning hte cluster with highest acheived Jaccard index. Afterwards, the Jaccard indices are calculated, comparing cluster members with complete and leaky data, for each cluster.

References

Anja Jochmann, Luca Artusio, Hoda Sharifian, Angela Jamalzadeh, Louise J Fleming, Andrew Bush, Urs Frey, and Edgar Delgado-Eckert. Fluctuation-based clustering reveals phenotypes of patients with different asthma severity. ERJ open research, 6(2), 2020.

Edgar Delgado-Eckert, Oliver Fuchs, Nitin Kumar, Juha Pekkanen, Jean-Charles Dalphin, Josef Riedler, Roger Lauener, Michael Kabesch, Maciej Kupczyk, Sven-Erik Dahlen, et al. Functional phenotypes determined by fluctuation-based clustering of lung function measurements in healthy and asthmatic cohort participants. Thorax, 73(2):107–115, 2018.

Examples

list <- patient_list(
"https://raw.githubusercontent.com/MrMaximumMax/FBCanalysis/master/demo/phys/data.csv",
GitHub = TRUE)
#> 
#> 
#>   Patient PEF FEV1     date
#> 1  test_a   1   33 01.01.21
#> 2  test_a   8   42 02.01.21
#> 3  test_a  18   35 03.01.21
#> 4  test_a  19   47 04.01.21
#> 5  test_a  19   37 05.01.21
#> 
#> 
#>       V1  V2   V3   V4
#>  Patient PEF FEV1 date
#>        1   2    3    4
#> 
#> Which column represents the Patient_ID (please indicate number)?: 
#> Which column represents the time (please indicate number)?: 
#> 
#> 
#>  Day:Month:Year 	 (1) 
#>  Month:Day:Year 	 (2) 
#>  Year:Month:Day 	 (3) 
#>  Year:Day:Month 	 (4) 
#>  Day:Month:Year:Hour:Min (5) 
#>  Month:Day:Year:Hour:Min (6) 
#>  Year:Month:Day:Hour:Min (7) 
#>  Year:Day:Month:Hour:Min (8) 
#> 
#> What's the time format (please indicate number)?: 
#> 
#> 
#>  Twice daily 	 (1) 
#>  Daily 		 (2) 
#>  Twice a week 	 (3) 
#>  Weekly 	 (4) 
#>  Twice a month 	 (5) 
#>  Monthly 	 (6) 
#>  Twice a quarter (7) 
#>  Quarterly 	 (8) 
#> 
#> What's the sampling frequency (please indicate number)?: 
#> 
#> 
#>  Sample missing values from First quartile 	 (1) 
#>  Sample missing values from Third quartile 	 (2) 
#> 
#>  Sample missing values from 5th percentile 	 (3) 
#>  Sample missing values from 95th percentile 	 (4) 
#> 
#>  Interpolate missing values and apply L1 Regularization / Lasso Regression 		 (5) 
#>  Interpolate missing values and apply L2 Regulariztation / Ridge Regression 		 (6) 
#>  Interpolate missing values and apply Elastic Net on Regression and Regularization 	 (7) 
#> 
#>  Interpolate missing values by Linear Spline 	 (8) 
#>  Interpolate missing values by Cubic C2 Spline 	 (9) 
#> 
#> Select measure for filling up NA values: 
#> Error in `[.data.frame`(raw.data, , "Patient_ID"): undefined columns selected
#Sampling frequency is supposed to be daily
output <- sim_jaccard_global(list, "PEF", 0.05, 10, "hierarchical", 2, 1000)
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Error in plist[[i]]: object of type 'builtin' is not subsettable