Observe specific cluster for overview and p-values

Observe specific cluster for overview and p-values

enr_obs_clust(ts.dat, enrich, clustno)

Arguments

ts.dat

Processed data frame storing time series data and cluster assignments (also see function: add_clust2ts)

enrich

Processed data frame storing enrichment data and cluster assignments (also see function: add_clust2enrich)

clustno

Cluster number of interest

Value

Terminal output presenting summary of time series and enrichment data with corresponding p-values

Examples

list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#> Error in patient_list(".../ts_demofiles1"): could not find function "patient_list"
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
clustering <- clust_matrix(matrix, method = "kmeans", nclust = 3)
#> Error in clust_matrix(matrix, method = "kmeans", nclust = 3): could not find function "clust_matrix"
enr <- add_enrich(list, '.../enrichment_dat.csv') #file can be drawn from GitHub demo files
#> Error in add_enrich(list, ".../enrichment_dat.csv"): could not find function "add_enrich"
enr <- add_clust2enrich(enr, clustering)
#> Error in add_clust2enrich(enr, clustering): could not find function "add_clust2enrich"
ts <- add_clust2ts(list, clustering)
#> Error in add_clust2ts(list, clustering): could not find function "add_clust2ts"
enr_obs_clust(ts, enr, 1)
#> Error in enr_obs_clust(ts, enr, 1): could not find function "enr_obs_clust"