Determine pair of maximum fluctuation difference in a list storing time series data

Determine pair of maximum fluctuation difference in a list storing time series data

max_fluc(plist, parameter)

Arguments

plist

List storing patient time series data (also see function: patient_list(path))

parameter

Parameter of interest from time series data list

Value

Console output with Patient_ID pair, corresponding Earth Mover's Distance and visualized boxplot of both time series data distributions

Examples

list <- patient_list('.../ts_demofiles1') #Just folder; files can be pulled from GitHub demo files
#> Error in patient_list(".../ts_demofiles1"): could not find function "patient_list"
#(https://github.com/MrMaximumMax/FBCanalysis/tree/master/demo_and_testfiles/ts_demofiles1)
#Sampling frequency is supposed to be daily
matrix <- emd_matrix(list, "FEV1")
#> Error in emd_matrix(list, "FEV1"): could not find function "emd_matrix"
max_fluc(list, "PEF")
#> Error in max_fluc(list, "PEF"): could not find function "max_fluc"